feat: Add ripgrep installation script

+++ neovim/install.sh
@@ -9,20 +9,24 @@ source "$DOTFILES_ROOT/lib/logger.sh"
 source "$DOTFILES_ROOT/lib/symlink.sh"
 
 info 'Installing Neovim'
-if ! command -v brew > /dev/null; then
-  user 'Homebrew not found. Would you like to install Homebrew first? (y/n)'
-  read -r -p '> ' install_brew
-  if [[ "$install_brew" =~ ^[Yy]$ ]]; then
-    source "$DOTFILES_ROOT/homebrew/install.sh"
-    source "$DOTFILES_ROOT/homebrew/path.zsh"
-  else
-    fail 'Homebrew is required to install Neovim.'
+if ! command -v nvim > /dev/null; then
+  if ! command -v brew > /dev/null; then
+    user 'Homebrew not found. Would you like to install Homebrew first? (y/n)'
+    read -r -p '> ' install_brew
+    if [[ "$install_brew" =~ ^[Yy]$ ]]; then
+      source "$DOTFILES_ROOT/homebrew/install.sh"
+      source "$DOTFILES_ROOT/homebrew/path.zsh"
+    else
+      fail 'Homebrew is required to install Neovim.'
+    fi
   fi
+  brew install neovim
+  success 'Neovim installed'
+else
+  success 'Neovim is already installed'
 fi
-brew install neovim
-# Install dependencies.
-brew install ripgrep
+info 'Configuring Neovim'
 mkdir -p "${XDG_CONFIG_HOME:-$HOME}/.config"
 overwrite_all=false backup_all=false skip_all=false
 link_file "$DOTFILES_ROOT/neovim/nvim" "${XDG_CONFIG_HOME:-$HOME}/.config/nvim"
-success 'Neovim installed'
+success 'Neovim configured'
